backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pymongo==4.6.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
